更新migrations

+++ database/migrations/2017_08_16_144547_create_ims_yz_order_discount_table.php
@@ -12,19 +12,21 @@ class CreateImsYzOrderDiscountTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('yz_order_discount', function(Blueprint $table)
-		{
-			$table->increments('id');
-			$table->integer('uid')->default(0);
-			$table->integer('order_id');
-			$table->string('discount_code', 50)->default('')->comment('折扣代码');
-			$table->string('name', 100)->default('')->comment('名称');
-			$table->decimal('amount', 10)->default(0.00)->comment('金额');
-			$table->boolean('is_indirect')->default(0)->comment('间接计算出');
-			$table->integer('updated_at')->nullable();
-			$table->integer('created_at')->nullable();
-			$table->integer('deleted_at')->nullable();
-		});
+        if (!\Schema::hasTable('yz_order_discount')) {
+
+            Schema::create('yz_order_discount', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('uid')->default(0);
+                $table->integer('order_id');
+                $table->string('discount_code', 50)->default('')->comment('折扣代码');
+                $table->string('name', 100)->default('')->comment('名称');
+                $table->decimal('amount', 10)->default(0.00)->comment('金额');
+                $table->boolean('is_indirect')->default(0)->comment('间接计算出');
+                $table->integer('updated_at')->nullable();
+                $table->integer('created_at')->nullable();
+                $table->integer('deleted_at')->nullable();
+            });
+        }
 	}
 
 
@@ -35,7 +37,9 @@ class CreateImsYzOrderDiscountTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('ims_yz_order_discount');
+        if (\Schema::hasTable('yz_order_discount')) {
+            Schema::drop('ims_yz_order_discount');
+        }
 	}
 
 }
